No toolhead sensor software, split parameter file
Parameters are now located in a dedicated file (ercf_parameters.cfg) for easier updates.
This software version is dedicated to systems without any toolhead sensor, only relying on the encoder data from the ercf. As a result the system is simpler and faster, but there are some downsides for now.
Typically in situations where the ercf state is unknown (e.g. after a reboot), we can't know for sure if filament is below the extruder gear or not, resulting in redundant moves to eject filament.